Move permissions to presenter
