+++ memento/src/main/java/com/alexstyl/specialdates/upcoming/widget/today/RecentPeopleEventsView.kt
@@ -5,4 +5,5 @@ import com.alexstyl.specialdates.events.peopleevents.ContactEventsOnADate
 interface RecentPeopleEventsView {
     fun onNextDateLoaded(events: ContactEventsOnADate)
     fun onNoEventsFound()
+    fun askForContactPermission()
 }
